backend/requirements.txt
# PortfolioExpert Backend Dependencies
fastapi>=0.100.0
uvicorn>=0.23.0
# yfinance removed - using direct Yahoo Finance API
requests>=2.28.0
pandas>=2.0.0
numpy>=1.24.0
fredapi>=0.5.1
sqlalchemy>=2.0.0
pydantic>=2.0.0
python-dateutil>=2.8.2
aiosqlite>=0.19.0
python-dotenv>=1.0.0

